refactor (1:1채팅): 클래스 리팩토링 - 채팅 테스트 html은 추후에 삭제
